refactoring for mac (#16)
* refactoring for mac
* update lammps data

we are basically adding the concept of a script template,
and ensuring that each cloud backend generates the full bash
template (without a need for optparse) to get around those
bugs. In addition, I refactored to not save stuff to a temporary
directory, but rather to create a .scripts directory inside
of the experiment directory so all outputs are inspectable
and reproducible by default. This can only be disabled
with --cleanup added as a flag. In addition, since there
was a pattern of running functions with the experiment as
a variable, I created its own class proper.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>